+++ packages/html-fmt-vscode/CHANGELOG.md
@@ -8,7 +8,8 @@ to structure this file.
 
 ## [Unreleased]
 
-- Support for perl expressions inside tags inside attribute values
+- Fix: Support for perl expressions inside tags inside attribute values
+- Fix: Adding quotes for `value` attribute in `<input value="<TMPL_V something>">`
 
 ## [0.6.0]